[BE : DEV] : S07P12B309-210 : add deleteUser API

See merge request s07-webmobile1-sub2/S07P12B309!71

+++ backend/src/main/java/com/ssafy/api/service/ProjectsService.java
@@ -11,7 +11,7 @@ import com.ssafy.db.entity.Projects;
  */
 public interface ProjectsService {
 
-    Integer createProject(ProjectsCreatePostReq projectsCreatePostReq) throws UidNullException;
+    Long createProject(ProjectsCreatePostReq projectsCreatePostReq) throws UidNullException;
 
-    Projects getProject(int pid) throws ProjectNullException;
+    Projects getProject(Long pid) throws ProjectNullException;
 }
